# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "BPMT API's",
        "description": "Project management operations",
        "version": "1.0.0"
    },
    "host": "http://my-default-host.com",
    "basePath": "/api/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/v1/task": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Task creation",
                "description": "Returns created Task",
                "operationId": "create-task",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "taskName",
                        "in": "formData",
                        "description": "Task name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "description of the Task",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "startDate",
                        "in": "formData",
                        "description": "start datetime",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "endDate",
                        "in": "formData",
                        "description": "End Datetime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "estimatedHours",
                        "in": "formData",
                        "description": "Estimated Hours",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "takenHours",
                        "in": "formData",
                        "description": "Taken Hours",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "takenHours",
                        "in": "formData",
                        "description": "Taken Hours",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "Task Status('created', 'assigned', 'onhold', 'inprogress','completed', 'cancelled',' failed')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "priority",
                        "in": "formData",
                        "description": "Task Priority('critical', 'high', 'medium', 'low')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "milestone_id",
                        "in": "formData",
                        "description": "Id of the Milestone",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "task_assigned_to",
                        "in": "formData",
                        "description": "Id of the User",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "task_assigned_by",
                        "in": "formData",
                        "description": "Id of the User",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "dependent_task_id",
                        "in": "formData",
                        "description": "Dependent Task ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/v1/task/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Task details",
                "description": "Returns Task details",
                "operationId": "single Task",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Task Updation",
                "description": "Returns Updated Task",
                "operationId": "Update-task",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "id of the Task",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "taskName",
                        "in": "formData",
                        "description": "Task name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "description of the Task",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "startDate",
                        "in": "formData",
                        "description": "start datetime",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "endDate",
                        "in": "formData",
                        "description": "End Datetime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "estimatedHours",
                        "in": "formData",
                        "description": "Estimated Hours",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "takenHours",
                        "in": "formData",
                        "description": "Taken Hours",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "takenHours",
                        "in": "formData",
                        "description": "Taken Hours",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "Task Status('created', 'assigned', 'onhold', 'inprogress','completed', 'cancelled',' failed')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "priority",
                        "in": "formData",
                        "description": "Task Priority('critical', 'high', 'medium', 'low')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "milestone_id",
                        "in": "formData",
                        "description": "Id of the Milestone",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "task_assigned_to",
                        "in": "formData",
                        "description": "Id of the User",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "task_assigned_by",
                        "in": "formData",
                        "description": "Id of the User",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "dependent_task_id",
                        "in": "formData",
                        "description": "Dependent Task ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete a Task",
                "description": "Delete a Task",
                "operationId": "delete Task",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/v1/task/by-sprints/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Task List from sprints",
                "description": "Returns Task List from sprints",
                "operationId": "Task-List-from-sprints",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sprint Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}